feat: await import dummy proposals script
